# swagger.yml
openapi: 3.0.3
info:
  version: '1'
  title: Ghdata
  description: ''
servers:
  - description: dev
    url: localhost
paths:
...
  /garmin/oauthCallback:
    get:
      summary: Callback url for oauth server
      tags:
        - garmin-oauth-flow
      parameters:
        - in: path
          name: oauth_token
          required: true
          schema:
            type: string
        - in: path
          name: oauth_verifier
          required: true
          schema:
            type: string
      responses:
        302:
          description: User was authorized
  /garmin/authorize:
    get:
      summary: Start oauth flow
      tags:
        - garmin-oauth-flow
      parameters:
        - in: path
          name: device_id
          required: true
          schema:
            type: string
          description: Device id
      responses:
        302:
          description: Started oauth flow
...